Added REST API calls for Store & Recycler Transaction Requests, More methods added to DB client
Added RequestStorePurchase and RequestRecyclerOperation to the REST API controller.

Added game, inventory and world Item retrieval methods for the database client.

Added and improved database structs.

Added collectible data tables and types for storing Metaplex metadata in the database

GameServer is now instanced by the API controller to hold important encryption data and interact with the SDK with it.

Dataprotection provider and protector moved to GameServer class from the API controller.